requirements.txt
flask==2.3.3
python-pptx==0.6.21
werkzeug==2.3.7 